# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moskollektor Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/work-orders": {
      "post": {
        "tags": [
          "work-orders"
        ],
        "summary": "Post Work Order",
        "operationId": "post_work_order_api_v1_work_orders_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "work-orders"
        ],
        "summary": "Get Work Orders",
        "operationId": "get_work_orders_api_v1_work_orders_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "work_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Type"
            }
          },
          {
            "name": "assigned_to_current_user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Assigned To Current User"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-orders/{work_order_id}": {
      "get": {
        "tags": [
          "work-orders"
        ],
        "summary": "Get Work Order Detail",
        "operationId": "get_work_order_detail_api_v1_work_orders__work_order_id__get",
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Order Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-orders/{work_order_id}/actions": {
      "post": {
        "tags": [
          "work-orders"
        ],
        "summary": "Post Work Order Action",
        "operationId": "post_work_order_action_api_v1_work_orders__work_order_id__actions_post",
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Order Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkOrderActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/engineers": {
      "get": {
        "tags": [
          "work-orders"
        ],
        "summary": "Get Engineers",
        "operationId": "get_engineers_api_v1_engineers_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineerListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/work-orders/{work_order_id}/engineer-candidates": {
      "get": {
        "tags": [
          "work-orders"
        ],
        "summary": "Get Work Order Engineer Candidates",
        "operationId": "get_work_order_engineer_candidates_api_v1_work_orders__work_order_id__engineer_candidates_get",
        "parameters": [
          {
            "name": "work_order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Work Order Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EngineerListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "work-orders"
        ],
        "summary": "Get Notifications",
        "operationId": "get_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "tags": [
          "work-orders"
        ],
        "summary": "Read Notification",
        "operationId": "read_notification_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Notification Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkOrderRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "source_risk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Risk Id"
          },
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "target_entity_type": {
            "type": "string",
            "title": "Target Entity Type"
          },
          "target_entity_id": {
            "type": "string",
            "title": "Target Entity Id"
          },
          "work_type": {
            "type": "string",
            "title": "Work Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "symptoms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symptoms"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 8
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "client_occurred_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Occurred At"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "facility_id",
          "target_entity_type",
          "target_entity_id",
          "work_type",
          "priority",
          "due_at",
          "description"
        ],
        "title": "CreateWorkOrderRequest",
        "description": "Backward-compatible draft creation request."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WorkOrderActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "edit",
              "submit",
              "start_triage",
              "request_clarification",
              "resubmit_clarification",
              "finalize_priority",
              "assign",
              "reassign",
              "accept",
              "decline",
              "mark_en_route",
              "start_work",
              "wait_access",
              "wait_parts",
              "resume_work",
              "submit_result",
              "start_verification",
              "return_for_rework",
              "close",
              "cancel",
              "override"
            ],
            "title": "Action"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "client_occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Client Occurred At"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "action",
          "expected_version",
          "idempotency_key",
          "client_occurred_at",
          "payload"
        ],
        "title": "WorkOrderActionRequest"
      },
      "WorkOrderActionResponse": {
        "properties": {
          "work_order": {
            "$ref": "#/components/schemas/WorkOrderOut"
          },
          "applied_action": {
            "type": "string",
            "enum": [
              "edit",
              "submit",
              "start_triage",
              "request_clarification",
              "resubmit_clarification",
              "finalize_priority",
              "assign",
              "reassign",
              "accept",
              "decline",
              "mark_en_route",
              "start_work",
              "wait_access",
              "wait_parts",
              "resume_work",
              "submit_result",
              "start_verification",
              "return_for_rework",
              "close",
              "cancel",
              "override"
            ],
            "title": "Applied Action"
          },
          "audit_event_id": {
            "type": "string",
            "title": "Audit Event Id"
          }
        },
        "type": "object",
        "required": [
          "work_order",
          "applied_action",
          "audit_event_id"
        ],
        "title": "WorkOrderActionResponse"
      },
...
      "WorkOrderListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WorkOrderOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/WorkOrderListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "WorkOrderListEnvelope"
      },
...
    }
  }
}